+++ corelib/src/libs/SireMove/openmmfrenergyst.h
@@ -122,6 +122,9 @@ namespace SireMove
         bool getMCBarostat(void);
         void setMCBarostat(bool);
 
+        bool getMCBarostatMembrane(void);
+        void setMCBarostatMembrane(bool);
+
         void setMCBarostatFrequency(int);
         int getMCBarostatFrequency(void);
 
@@ -247,6 +250,7 @@ namespace SireMove
         double Andersen_frequency;
 
         bool MCBarostat_flag;
+        bool MCBarostat_membrane_flag;
         int MCBarostat_frequency;
 
         QString ConstraintType;