+++ corelib/src/libs/SireMove/openmmmdintegrator.cpp
@@ -124,7 +124,7 @@ QDataStream &operator<<(QDataStream &ds, const OpenMMMDIntegrator &velver)
 
     sds << velver.frequent_save_velocities << velver.molgroup << velver.Integrator_type << velver.friction
         << velver.CutoffType << velver.cutoff_distance << velver.field_dielectric << velver.tolerance_ewald_pme
-        << velver.Andersen_flag << velver.Andersen_frequency << velver.MCBarostat_flag << velver.MCBarostat_frequency
+        << velver.Andersen_flag << velver.Andersen_frequency << velver.MCBarostat_flag << velver.MCBarostat_membrane_flag << velver.MCBarostat_frequency
         << velver.ConstraintType << velver.Pressure << velver.Temperature << velver.platform_type
         << velver.Restraint_flag << velver.CMMremoval_frequency << velver.buffer_frequency << velver.device_index
         << velver.LJ_dispersion << velver.precision << velver.integration_tol << velver.timeskip
@@ -147,7 +147,7 @@ QDataStream &operator>>(QDataStream &ds, OpenMMMDIntegrator &velver)
 
         sds >> velver.frequent_save_velocities >> velver.molgroup >> velver.Integrator_type >> velver.friction >>
             velver.CutoffType >> velver.cutoff_distance >> velver.field_dielectric >> velver.tolerance_ewald_pme >>
-            velver.Andersen_flag >> velver.Andersen_frequency >> velver.MCBarostat_flag >>
+            velver.Andersen_flag >> velver.Andersen_frequency >> velver.MCBarostat_flag >> velver.MCBarostat_membrane_flag >>
             velver.MCBarostat_frequency >> velver.ConstraintType >> velver.Pressure >> velver.Temperature >>
             velver.platform_type >> velver.Restraint_flag >> velver.CMMremoval_frequency >> velver.buffer_frequency >>
             velver.device_index >> velver.LJ_dispersion >> velver.precision >> velver.integration_tol >>
@@ -195,7 +195,7 @@ OpenMMMDIntegrator::OpenMMMDIntegrator(bool frequent_save)
       molgroup(MoleculeGroup()), openmm_system(0), openmm_context(0), isSystemInitialised(false),
       isContextInitialised(false), Integrator_type("leapfrogverlet"), friction(1.0 / picosecond),
       CutoffType("nocutoff"), cutoff_distance(1.0 * nanometer), field_dielectric(78.3), tolerance_ewald_pme(0.0001),
-      Andersen_flag(false), Andersen_frequency(90.0), MCBarostat_flag(false), MCBarostat_frequency(25),
+      Andersen_flag(false), Andersen_frequency(90.0), MCBarostat_flag(false), MCBarostat_membrane_flag(false), MCBarostat_frequency(25),
       ConstraintType("none"), Pressure(1.0 * bar), Temperature(300.0 * kelvin), platform_type("Reference"),
       Restraint_flag(false), CMMremoval_frequency(0), buffer_frequency(0), device_index("0"), LJ_dispersion(true),
       precision("single"), reinetialise_context(false), integration_tol(0.001), timeskip(0.0 * picosecond),
@@ -211,7 +211,7 @@ OpenMMMDIntegrator::OpenMMMDIntegrator(const MoleculeGroup &molecule_group, bool
       molgroup(molecule_group), openmm_system(0), openmm_context(0), isSystemInitialised(false),
       isContextInitialised(false), Integrator_type("leapfrogverlet"), friction(1.0 / picosecond),
       CutoffType("nocutoff"), cutoff_distance(1.0 * nanometer), field_dielectric(78.3), tolerance_ewald_pme(0.0001),
-      Andersen_flag(false), Andersen_frequency(90.0), MCBarostat_flag(false), MCBarostat_frequency(25),
+      Andersen_flag(false), Andersen_frequency(90.0), MCBarostat_flag(false), MCBarostat_membrane_flag(false), MCBarostat_frequency(25),
       ConstraintType("none"), Pressure(1.0 * bar), Temperature(300.0 * kelvin), platform_type("Reference"),
       Restraint_flag(false), CMMremoval_frequency(0), buffer_frequency(0), device_index("0"), LJ_dispersion(true),
       precision("single"), reinetialise_context(false), integration_tol(0.001), timeskip(0.0 * picosecond),
@@ -229,7 +229,7 @@ OpenMMMDIntegrator::OpenMMMDIntegrator(const OpenMMMDIntegrator &other)
       Integrator_type(other.Integrator_type), friction(other.friction), CutoffType(other.CutoffType),
       cutoff_distance(other.cutoff_distance), field_dielectric(other.field_dielectric),
       tolerance_ewald_pme(other.tolerance_ewald_pme), Andersen_flag(other.Andersen_flag),
-      Andersen_frequency(other.Andersen_frequency), MCBarostat_flag(other.MCBarostat_flag),
+      Andersen_frequency(other.Andersen_frequency), MCBarostat_flag(other.MCBarostat_flag), MCBarostat_membrane_flag(other.MCBarostat_membrane_flag),
       MCBarostat_frequency(other.MCBarostat_frequency), ConstraintType(other.ConstraintType), Pressure(other.Pressure),
       Temperature(other.Temperature), platform_type(other.platform_type), Restraint_flag(other.Restraint_flag),
       CMMremoval_frequency(other.CMMremoval_frequency), buffer_frequency(other.buffer_frequency),
@@ -264,6 +264,7 @@ OpenMMMDIntegrator &OpenMMMDIntegrator::operator=(const OpenMMMDIntegrator &othe
     Andersen_flag = other.Andersen_flag;
     Andersen_frequency = other.Andersen_frequency;
     MCBarostat_flag = other.MCBarostat_flag;
+    MCBarostat_membrane_flag = other.MCBarostat_membrane_flag;
     MCBarostat_frequency = other.MCBarostat_frequency;
     ConstraintType = other.ConstraintType;
     Pressure = other.Pressure;
@@ -291,7 +292,7 @@ bool OpenMMMDIntegrator::operator==(const OpenMMMDIntegrator &other) const
            isSystemInitialised == other.isSystemInitialised and isContextInitialised == other.isContextInitialised and
            CutoffType == other.CutoffType and cutoff_distance == other.cutoff_distance and
            field_dielectric == other.field_dielectric and Andersen_flag == other.Andersen_flag and
-           Andersen_frequency == other.Andersen_frequency and MCBarostat_flag == other.MCBarostat_flag and
+           Andersen_frequency == other.Andersen_frequency and MCBarostat_flag == other.MCBarostat_flag and MCBarostat_membrane_flag == other.MCBarostat_membrane_flag and
            MCBarostat_frequency == other.MCBarostat_frequency and ConstraintType == other.ConstraintType and
            Pressure == other.Pressure and Temperature == other.Temperature and platform_type == other.platform_type and
            Restraint_flag == other.Restraint_flag and CMMremoval_frequency == other.CMMremoval_frequency and
@@ -487,9 +488,21 @@ void OpenMMMDIntegrator::initialise()
         const double converted_Temperature = convertTo(Temperature.value(), kelvin);
         const double converted_Pressure = convertTo(Pressure.value(), bar);
 
-        OpenMM::MonteCarloBarostat *barostat =
-            new OpenMM::MonteCarloBarostat(converted_Pressure, converted_Temperature, MCBarostat_frequency);
-        system_openmm->addForce(barostat);
+        if (MCBarostat_membrane_flag == true)
+        {
+            // Simple options for now: zero surface tension, XY isotropic, Z free
+            const double surface_Tension = 0;
+            OpenMM::MonteCarloMembraneBarostat::XYMode xymode = OpenMM::MonteCarloMembraneBarostat::XYIsotropic;
+            OpenMM::MonteCarloMembraneBarostat::ZMode zmode = OpenMM::MonteCarloMembraneBarostat::ZFree;
+            OpenMM::MonteCarloMembraneBarostat * barostat = new OpenMM::MonteCarloMembraneBarostat(converted_Pressure, surface_Tension, converted_Temperature, xymode, zmode, MCBarostat_frequency);
+            system_openmm->addForce(barostat);
+        }
+        else    // normal barostat
+        {
+            OpenMM::MonteCarloBarostat *barostat =
+                new OpenMM::MonteCarloBarostat(converted_Pressure, converted_Temperature, MCBarostat_frequency);
+            system_openmm->addForce(barostat);
+        }
 
         if (Debug)
         {
@@ -498,6 +511,10 @@ void OpenMMMDIntegrator::initialise()
             qDebug() << "Pressure = " << converted_Pressure << " bar\n";
             qDebug() << "Frequency every " << MCBarostat_frequency << " steps\n";
             qDebug() << "Lennard Jones Dispersion term is set to " << LJ_dispersion << "\n";
+            if (MCBarostat_membrane_flag)
+            {
+                qDebug() << "Membrane barostat, surface tension 0, XY isotropic, Z free\n";
+            }
         }
     }
 
@@ -1926,6 +1943,16 @@ void OpenMMMDIntegrator::setMCBarostat(bool MCBarostat)
 bool OpenMMMDIntegrator::getMCBarostat(void)
 {
     return MCBarostat_flag;
+}
+
+void OpenMMMDIntegrator::setMCBarostatMembrane(bool MCBarostatMembrane)
+{
+    MCBarostat_membrane_flag = MCBarostatMembrane;
+}
+
+bool OpenMMMDIntegrator::getMCBarostatMembrane(void)
+{
+    return MCBarostat_membrane_flag;
 }
 
 /** Get the Monte Carlo Barostat frequency in time speps */