+++ wrapper/Tools/OpenMMMD.py
@@ -268,6 +268,12 @@ barostat = Parameter(
     """Whether or not to use a barostat (needed for NPT simulation).""",
 )
 
+barostat_membrane = Parameter(
+    "membrane barostat",
+    False,
+    """Whether the barostat is a membrane barostat (needed for membrane simulation).""",
+)
+
 andersen_frequency = Parameter(
     "andersen frequency", 10.0, """Collision frequency in units of (1/ps)"""
 )
@@ -824,6 +830,8 @@ def setupMoves(system, debug_seed, GPUS):
     if barostat.val:
         Integrator_OpenMM.setPressure(pressure.val)
         Integrator_OpenMM.setMCBarostat(barostat.val)
+        if barostat_membrane.val:
+            Integrator_OpenMM.setMCBarostatMembrane(barostat_membrane.val)
         Integrator_OpenMM.setMCBarostatFrequency(barostat_frequency.val)
 
     # print Integrator_OpenMM.getDeviceIndex()
@@ -2145,6 +2153,8 @@ def setupMovesFreeEnergy(system, debug_seed, gpu_idx, lam_val):
     if barostat.val:
         Integrator_OpenMM.setPressure(pressure.val)
         Integrator_OpenMM.setMCBarostat(barostat.val)
+        if barostat_membrane.val:
+            Integrator_OpenMM.setMCBarostatMembrane(barostat_membrane.val)
         Integrator_OpenMM.setMCBarostatFrequency(barostat_frequency.val)
 
     # Choose a random seed for Sire if a debugging seed hasn't been set.