+++ doc/source/changelog.rst
@@ -42,6 +42,8 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
 * Fix the ``hasForceSpecificEquation`` function in the ``LambdaLever`` class so that it returns true if
   there is a default equation for the force.
 
+* Added support for the ``OpenMM`` ``MonteCarloMembraneBarostat`` to ``SOMD``.
+
 `2024.3.1 <https://github.com/openbiosim/sire/compare/2024.3.0...2024.3.1>`__ - December 2024
 --------------------------------------------------------------------------------------------